Add if clauses to each chunk to avoid unnecessary repetition

+++ start.sh
@@ -26,7 +26,9 @@ echo "$greeting"
 # Silence virtual console default beep
 #######################################
 
-silencer=$(cat << EOF
+# Check if the service file exists
+if ! sudo systemctl list-unit-files --type=service | grep -q 'silence-console.service'; then
+    silencer=$(cat << EOF
 [Unit]
 Description=Silence virtual console default beep
 
@@ -40,15 +42,21 @@ ExecStart=/usr/bin/setterm -blength 0
 [Install]
 WantedBy=multi-user.target
 EOF
-)
+    )
 
-sudo tee -a "/etc/systemd/system/silence-console.service" > /dev/null <<EOT
+    sudo tee "/etc/systemd/system/silence-console.service" > /dev/null <<EOT
 $silencer
 EOT
 
-sudo systemctl daemon-reload
-sudo systemctl enable silence-console
-sudo systemctl start silence-console
+    sudo systemctl daemon-reload
+    sudo systemctl enable silence-console
+    sudo systemctl start silence-console
+
+    echo "Service silence-console created and started."
+else
+    echo "Service silence-console already exists. Skipping."
+fi
+
 #######################################
 
 
@@ -56,10 +64,18 @@ sudo systemctl start silence-console
 # Pull dotfiles
 #######################################
 
-git clone --bare https://github.com/"$USER"/.dotfiles.git "$HOME"/.dotfiles
-alias dotfiles='/usr/bin/git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME'
-dotfiles config --local status.showUntrackedFiles no
-dotfiles checkout
+dotfiles_dir="$HOME/.dotfiles"
+
+# Check if the ~/.dotfiles folder already exists
+if [ ! -d "$dotfiles_dir" ]; then
+    git clone --bare https://github.com/"$USER"/.dotfiles.git "$HOME"/.dotfiles
+    alias dotfiles='/usr/bin/git --git-dir=$HOME/.dotfiles/ --work-tree=$HOME'
+    dotfiles config --local status.showUntrackedFiles no
+    dotfiles checkout
+    echo "Dotfiles setup completed."
+else
+    echo "Dotfiles folder already exists. Skipping setup."
+fi
 
 #######################################
 
@@ -83,6 +99,8 @@ network-manager
 build-essential
 libpam0g-dev
 libxcb-xkb-dev
+ca-certificates 
+gnupg
 alsa-utils
 pulseaudio
 pavucontrol
@@ -91,12 +109,27 @@ EOF
 
 sudo apt update && sudo apt install -y "$packages"
 chsh -s /usr/bin/zsh
-sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
-       https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
 sudo update-alternatives --set editor /usr/bin/nvim
 sudo update-alternatives --set x-terminal-emulator /usr/bin/xterm
+
 #######################################
 
+
+#######################################
+# Install vim-plug
+#######################################
+
+target_path="${XDG_DATA_HOME:-$HOME/.local/share}/nvim/site/autoload/plug.vim"
+if [ ! -f "$target_path" ]; then
+    sh -c 'curl -fLo "$target_path" --create-dirs \
+           https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
+
+    echo "vim-plug installation completed."
+else
+    echo "vim-plug is already downloaded. Skipping installation."
+fi
+
+#######################################
 
 #######################################
 # Update Grub
@@ -112,13 +145,18 @@ sudo update-grub
 # Set up Ly
 #######################################
 
-git clone --recurse-submodules https://github.com/"$USER"/ly.git &&\
-sh -c cd ly \
- make \
- sudo make install installsystemd \
- sudo systemctl enable ly.service \
- sudo systemctl set-default graphical.target
-rm -rf ly
+if ! sudo systemctl list-unit-files --type=service | grep -q 'ly.service'; then
+  git clone --recurse-submodules https://github.com/"$USER"/ly.git &&\
+  sh -c cd ly \
+   make \
+   sudo make install installsystemd \
+   sudo systemctl enable ly.service \
+   sudo systemctl set-default graphical.target
+  rm -rf ly
+  echo "Service ly created and started."
+else
+  echo "Service ly already exists. Skipping."
+fi
 
 #######################################
 
@@ -127,9 +165,14 @@ rm -rf ly
 # Install Miniconda
 #######################################
 
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/install_miniconda.sh
-sh ~/install_miniconda.sh -b -u -p ~/.miniconda3
-rm -rf ~/install_miniconda.sh
+if ! type conda; then
+  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/install_miniconda.sh
+  sh ~/install_miniconda.sh -b -u -p ~/.miniconda3
+  rm -rf ~/install_miniconda.sh
+  echo "Conda installed."
+else
+  echo "Conda already installed. Skipping."
+fi
 
 #######################################
 
@@ -138,11 +181,16 @@ rm -rf ~/install_miniconda.sh
 # Install Brave browser
 #######################################
 
-sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
-echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
-sudo apt update
-sudo apt install -y brave-browser
-sudo update-alternatives --set x-www-browser /usr/bin/brave-browser-stable
+if ! type brave-browser; then
+  sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
+  echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
+  sudo apt update
+  sudo apt install -y brave-browser
+  sudo update-alternatives --set x-www-browser /usr/bin/brave-browser-stable
+  echo "Brave browser installed."
+else
+  echo "Brave browser already installed. Skipping."
+fi
 
 #######################################
 
@@ -151,11 +199,16 @@ sudo update-alternatives --set x-www-browser /usr/bin/brave-browser-stable
 # Install Github CLI
 #######################################
 
-curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg \
-&& sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg \
-&& echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null \
-&& sudo apt update \
-&& sudo apt install gh -y
+if ! type gh; then
+  curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg \
+  && echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null \
+  && sudo apt update \
+  && sudo apt install gh -y
+  echo "Github CLI installed."
+else
+  echo "Github CLI already installed. Skipping."
+fi
 
 #######################################
 
@@ -164,19 +217,23 @@ curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo 
 # Install Docker
 #######################################
 
-sudo apt install -y ca-certificates curl gnupg
-sudo install -m 0755 -d /etc/apt/keyrings
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-sudo chmod a+r /etc/apt/keyrings/docker.gpg
-echo \
-  "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
-  "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-sudo apt update
-sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
-sudo service docker start
-sudo docker run hello-world
-sudo usermod -aG docker "$USER"
+if ! type docker; then
+  sudo install -m 0755 -d /etc/apt/keyrings
+  curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+  sudo chmod a+r /etc/apt/keyrings/docker.gpg
+  echo \
+    "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
+    "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
+    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+  sudo apt update
+  sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+  sudo service docker start
+  sudo docker run hello-world
+  sudo usermod -aG docker "$USER"
+  echo "Docker installed."
+else
+  echo "Docker already installed. Skipping."
+fi
 
 #######################################
